Extract interface for the HttpClientFactory.
