+++ library/src/main/java/com/novoda/downloadmanager/HttpClientFactory.java
@@ -1,33 +1,7 @@
 package com.novoda.downloadmanager;
 
+public interface HttpClientFactory {
 
-import java.util.concurrent.TimeUnit;
+    HttpClient create();
 
-import okhttp3.OkHttpClient;
-
-final class HttpClientFactory {
-
-    private static final int TIMEOUT = 5;
-
-    private HttpClientFactory() {
-        // non-instantiable class
-    }
-
-    public static HttpClient getInstance() {
-        return LazySingleton.INSTANCE;
-    }
-
-    private static class LazySingleton {
-
-        private static final HttpClient INSTANCE = createInstance();
-
-        private static HttpClient createInstance() {
-            OkHttpClient okHttpClient = new OkHttpClient.Builder()
-                    .connectTimeout(TIMEOUT, TimeUnit.SECONDS)
-                    .writeTimeout(TIMEOUT, TimeUnit.SECONDS)
-                    .readTimeout(TIMEOUT, TimeUnit.SECONDS)
-                    .build();
-            return new WrappedOkHttpClient(okHttpClient);
-        }
-    }
 }
